Move private API elements out of pk.h
